feat: split integration tabs into pages (#571)
* feat: split syncs and integrations into separate pages

* chore: update links to syncs and credentials pages

* fix: sidebar menu active state

---------
